feat: add speakers tab
