+++ lib/schedule/models/models.dart
@@ -2,6 +2,5 @@ export 'activity.dart';
 export 'event.dart';
 export 'location.dart';
 export 'schedule.dart';
-export 'speaker.dart';
 export 'talk.dart';
 export 'workshop.dart';
