+++ lib/speaker_details/speaker_details.dart
@@ -1,0 +1,1 @@
+export 'view/view.dart';
