+++ lib/speaker_details/view/view.dart
@@ -1,0 +1,1 @@
+export 'speaker_details_page.dart';
