+++ lib/speakers/data/speakers.dart
@@ -1,0 +1,258 @@
+import 'package:flutter_and_friends/speakers/speakers.dart';
+
+const speakers = [
+  albertWolszon,
+  annaLeushchenko,
+  eliasElfarri,
+  enzoConty,
+  esraKadah,
+  filipHracek,
+  ingvildSandstad,
+  ivannaKacevica,
+  jamieKerber,
+  jochumVanDerPloeg,
+  luanNico,
+  lucasJosefiak,
+  mangirdasKazlauskas,
+  martinOlsson,
+  mateuszWojtczak,
+  morganHunt,
+  moritzTheis,
+  poojaBhaumik,
+  remiRousselet,
+  renanAraujo,
+  roaaKhaddam,
+  rudolfsAgrisStilve,
+  sandraLundh,
+  sashaDenisov,
+  slavaEgorov,
+];
+
+const luanNico = Speaker(
+  name: 'Luan Nico',
+  title: 'Maintainer at Blue Fire',
+  bio:
+      '''Original creator of the Flame Engine, one of its passionate maintainers and currently working to make Flame 3D more ergonomic to use. Zealous advocate for OSS, Physics, and trailing commas.''',
+  avatar: 'assets/speakers/luan_nico.jpg',
+  twitter: 'luanpotter27',
+);
+
+const mateuszWojtczak = Speaker(
+  name: 'Mateusz Wojtczak',
+  title: 'Head of Mobile at LeanCode',
+  bio: '''
+Led the first full Flutter retail banking app - Credit Agricole
+Flutter in Enterprise preacher
+Patrol maintainer
+''',
+  avatar: 'assets/speakers/mateusz_wojtczak.jpg',
+  twitter: 'matiwojt',
+);
+
+const renanAraujo = Speaker(
+  name: 'Renan Araujo',
+  title: 'Software Engineer at Superlist',
+  bio:
+      '''Loves to use programming to create user interfaces. Worked previously with the web and have been in love with Flutter since beta. Created and helped to maintain some open-source projects.''',
+  avatar: 'assets/speakers/renan_araujo.jpg',
+  twitter: 'reNotANumber',
+);
+
+const sashaDenisov = Speaker(
+  name: 'Sasha Denisov',
+  title: 'GDE in AI, Flutter and Firebase, @EPAM',
+  bio:
+      '''Sasha is a software developer with over 20 years in multiple technologies and domains. A tech lead, architect, and mobile solutions expert, he specializes in AI integrations for mobile apps, focusing on on-device AI. He excels in Flutter, Firebase, and generative AI for robust, scalable apps. Sasha is Chief Software Engineer and Head of Flutter Discipline at EPAM, GDE for AI, Firebase, Flutter, Dart, and co-organizes Flutter Berlin Community''',
+  avatar: 'assets/speakers/sasha_denisov.jpeg',
+  twitter: 'ShuregDenisov',
+);
+
+const jochumVanDerPloeg = Speaker(
+  name: 'Jochum van der Ploeg',
+  title: 'Software Engineer @ FlutterFlow',
+  bio: '''
+I'm a Dutch software engineer, working at FlutterFlow in the day and maintain and build the 🔥 Flame game engine by night.
+
+My hobbies are:
+- Annoying Lukas
+- Building games with Flame
+- Making people hate the Dart language
+
+If you are looking for a fun time, swipe right and I can guarantee you won't regret it (you have to join my workshop otherwise you will actually regret it)
+''',
+  avatar: 'assets/speakers/jochum_van_der_ploeg.jpeg',
+  twitter: 'wolfenrain',
+);
+
+const sandraLundh = Speaker(
+  name: 'Sandra Lundh',
+  title: 'Software Developer @ Publicis Sapient',
+  bio: '''
+She has been developing apps in Flutter since the first stable version and have not stopped since!
+Accessibly and scalable design is important to her as she wants to build apps that everyone can use.
+''',
+  avatar: 'assets/speakers/sandra_lundh.jpeg',
+);
+
+const morganHunt = Speaker(
+  name: 'Morgan Hunt',
+  title: 'Senior Software Engineer',
+  bio:
+      '''Morgan has been building with Flutter for over 5 years, creating tools, libraries, and mobile apps that aim to make developers' lives easier. He's all about clean architecture, great DX, and finding clever ways to push Flutter further. When he's not coding, he's probably reading sci-fi, tinkering with a side project, or replaying a Zelda game for the millionth time.''',
+  avatar: 'assets/speakers/morgan_hunt.jpeg',
+  twitter: 'mrgnhnt96_dev',
+);
+
+const remiRousselet = Speaker(
+  name: 'Remi Rousselet',
+  title: 'Software Engineer',
+  bio: 'Author of Riverpod/Freezed/flutter_hooks.',
+  avatar: 'assets/speakers/remi_rousselet.jpeg',
+  twitter: 'remi_rousselet',
+);
+
+const annaLeushchenko = Speaker(
+  name: 'Anna Leushchenko',
+  title: 'Senior Staff Engineer / Engineering Manager',
+  bio:
+      '''Anna Leushchenko is a Google Developer Expert in Dart and Flutter, and a Senior Staff Engineer / Engineering Manager at UK-based fintech company Tide. With over a decade in mobile development, she leads cross-functional teams, delivers scalable mobile solutions, drives engineering excellence, and shares her expertise at global tech conferences.''',
+  avatar: 'assets/speakers/anna_leushchenko.png',
+  twitter: 'AnnaLeushchenko',
+);
+
+const jamieKerber = Speaker(
+  name: 'Jamie Kerber',
+  title: 'Senior Software Engineer, consulting @ Very Good Ventures',
+  bio:
+      '''Jamie Kerber is a Senior Software Engineer, currently consulting at Very Good Ventures. With over 12 years of experience building scalable cloud systems and game engines, and 6 years in Flutter/Dart, she builds and designs delicious apps that sure to perform well and look great. She's also a Flutter OSS contributor focusing on iOS widget fidelity.''',
+  avatar: 'assets/speakers/jamie_kerber.jpeg',
+  twitter: 'kerberjg',
+);
+
+const ingvildSandstad = Speaker(
+  name: 'Ingvild Sandstad',
+  title: 'Senior developer @ Giant Leap Technologies',
+  bio:
+      '''Ingvild used to live an analog, acoustic life as a classical musician until about five years ago when the world stopped and she suddenly had no concerts to play. She opened her computer instead, discovered coding and hasn't looked back. She has been developing Flutter apps for four years now.''',
+  avatar: 'assets/speakers/ingvild_sandstad.jpeg',
+);
+
+const eliasElfarri = Speaker(
+  name: 'Elias Elfarri',
+  title: 'App Lead @ Fink',
+  bio:
+      '''Elias is the main organizer of Flutter Oslo, Norway's largest Flutter community. He helped bring Flutter into the Norwegian oil industry, leading a mobile team at InlineX. Elias is also an active open-source contributor and comments on Flutter-related matters in the Norwegian tech newspaper Kode24.''',
+  avatar: 'assets/speakers/elias_elfarri.jpeg',
+);
+
+const rudolfsAgrisStilve = Speaker(
+  name: 'Rūdolfs Agris Stilve',
+  title: 'Software Engineer @ HackMotion',
+  bio:
+      '''Rūdolfs is a software engineer at HackMotion, building a market-leading virtual coach for golfers, powered by Flutter. With a background in game development, embedded software, and 4 years of experience in Flutter, his aim is to make the most delightful and reliable product possible.''',
+  avatar: 'assets/speakers/rudolfs_stilve.jpeg',
+);
+
+const martinOlsson = Speaker(
+  name: 'Martin Olsson',
+  title: 'Lead App & UI Engineer @ Candela',
+  bio:
+      '''I'm a sailor and a coder and I develop and design complex things that are easy to use. I draw from experience in games for children, like Mulle Meck Bygger Båtar, learning platforms like Unikum, life-saving systems like the Nectarine Health elderly care network, and life-changing experiences at The Story Lab. I've used Flutter since before 1.0, and I'm now using it to build interfaces and companion apps for foiling boats and ferries.''',
+  avatar: 'assets/speakers/martin_olsson.jpeg',
+);
+
+const lucasJosefiak = Speaker(
+  name: 'Lucas Josefiak',
+  title: 'Co-Founder & CEO',
+  bio: 'Co-Founder & CEO of Widgetbook',
+  avatar: 'assets/speakers/lucas_josefiak.jpeg',
+  twitter: 'LucasJosefiak',
+);
+
+const enzoConty = Speaker(
+  name: 'Enzo Conty',
+  title: 'CTO @ Ezymob',
+  bio:
+      '''Enzo Conty currently serves as the CTO of Ezymob, applying his extensive skills as a Sr. Mobile SWE honed over seven years of freelancing. Renowned for his expertise in Flutter, he has developed numerous innovative applications exploring new platforms such as Steam Deck, embeddable systems, drink distributors and even Steam games.Enzo also enriches the tech community being a Google Developer Expert (GDE) and organizing the GDG Bordeaux and Flutter Meetup Group Bordeaux.''',
+  avatar: 'assets/speakers/enzo_conty.jpeg',
+  twitter: 'EnzoConty',
+);
+
+const albertWolszon = Speaker(
+  name: 'Albert Wolszon',
+  title: 'Senior Flutter Developer',
+  bio:
+      '''Flutter GDE. Flutter developer since Flutter 1.0.0. Specializing in UI development and internationalization. Responsible for the development of the company's and a few big clients' design systems. Creator of a Flutter-dedicated TMS. After work flying as a private pilot.''',
+  avatar: 'assets/speakers/albert_wolszon.jpeg',
+  twitter: 'albert_wolszon',
+);
+
+const filipHracek = Speaker(
+  name: 'Filip Hracek',
+  title: 'Freelance OS & Game Developer',
+  bio:
+      '''Filip is a programming buff with formal training in journalism. He works on OSS projects and games in Flutter.''',
+  avatar: 'assets/speakers/filip_hracek.jpeg',
+  twitter: 'filiphracek',
+);
+
+const ivannaKacevica = Speaker(
+  name: 'Ivanna Kaceviča',
+  title: 'Lead Software Engineer @LMT',
+  bio:
+      '''Ivanna is an award-winning Flutter developer and a Lead Software Engineer at LMT, Latvia's first and largest mobile telecommunications operator. Over the past couple of years, she has won three coding scholarship competitions and several contests, securing the Grand Prize in both the Flutter Global Gamers Challenge and the Google Cloud NCAA March Madness Competition.''',
+  avatar: 'assets/speakers/ivanna_kacevica.png',
+);
+
+const slavaEgorov = Speaker(
+  name: 'Slava Egorov',
+  title: 'Senior Staff Software Engineer @ Google',
+  bio: 'I am an uber-TL for the Dart programming language.',
+  avatar: 'assets/speakers/slava_egorov.jpeg',
+  twitter: 'mraleph',
+);
+
+const mangirdasKazlauskas = Speaker(
+  name: 'Mangirdas Kazlauskas',
+  title: 'Mobile Tech Lead @ Billo',
+  bio:
+      '''An experienced product-centric software engineer with a passion for learning and sharing knowledge. As a Google Developer Expert for Flutter & Dart technologies, I excel at building cross-platform apps with Flutter, and I love teaching others what I know. In my free time, I enjoy creating educational content, or just being outside and playing disc golf.''',
+  avatar: 'assets/speakers/mangirdas_kazlauskas.jpeg',
+  twitter: 'mkobuolys',
+);
+
+const esraKadah = Speaker(
+  name: 'Esra Kadah',
+  title: 'Senior Flutter Engineer, Author, Community Lead @Antigua Mobile',
+  bio:
+      '''Esra Kadah is a Senior Flutter Developer at Antigua Mobile and co-author of Flutter Design Patterns & Best Practices. She leads community efforts like Flutter Community AI Circle, Flutteristas, and Flutter Berlin Meetup. Passionate about blending code, design, and human-centered AI, she builds mindful, meaningful apps and spaces where developers feel seen & connected.''',
+  avatar: 'assets/speakers/esra_kadah.jpeg',
+  twitter: 'esratech',
+);
+
+const roaaKhaddam = Speaker(
+  name: 'Roaa Khaddam',
+  title: 'Senior Software Engineer @ simpleclub, Flutter & Dart GDE',
+  bio:
+      '''Flutter Google Developer Expert and Senior Software Engineer at simpleclub, helping shape the future of education. Enjoying the art of building with code since 2017, and with Flutter since 2019. Passionate about animations, generative art, creative coding, content creation, and speaking at tech events.''',
+  avatar: 'assets/speakers/roaa_khaddam.jpeg',
+  twitter: 'roaakdm',
+);
+
+const poojaBhaumik = Speaker(
+  name: 'Pooja Bhaumik',
+  title: 'Developer Advocate, FlutterFlow',
+  bio:
+      '''Pooja Bhaumik, a Google Developer Expert for Flutter, Developer Advocate at FlutterFlow, author & creator of Flutter Essential Training, an exclusive Linkedin Learning course. She has spoken at over 20 conferences in the past five years, including in-person international events like Fluttercon NYC & Berlin, Flutter Conf LATAM, and F3 Prague. She is also passionate about mentoring, especially women in STEM, and has participated in numerous panels on the subject.''',
+  avatar: 'assets/speakers/pooja_bhaumik.png',
+  twitter: 'pooja_bhaumik',
+);
+
+const moritzTheis = Speaker(
+  name: 'Moritz Theis',
+  title: 'CEO at Snapp X',
+  bio:
+      '''Moritz is a GDE for Flutter & Dart, Organizer of Flutter Munich and the Founder of Snapp X.''',
+  avatar: 'assets/speakers/moritz_theis.jpeg',
+  twitter: 'MoritzTheis',
+);
