+++ lib/speakers/models/models.dart
@@ -1,0 +1,1 @@
+export 'speaker.dart';
