+++ lib/speakers/models/speaker.dart
