+++ lib/speakers/speakers.dart
@@ -1,0 +1,4 @@
+export 'data/data.dart';
+export 'models/models.dart';
+export 'view/view.dart';
+export 'widgets/widgets.dart';
