+++ lib/speakers/view/view.dart
@@ -1,0 +1,1 @@
+export 'speakers_page.dart';
