+++ lib/speakers/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'speaker_card.dart';
